requirements.txt
flask
python-dotenv
flask-jwt-extended
requests
